refactor: rename lib folder to internal
