+++ spec/unit/internal/broadcast.spec.ts
@@ -1,4 +1,4 @@
-import { broadcast } from "../../../src/lib/broadcast";
+import { broadcast } from "../../../src/internal/broadcast";
 
 // eslint-disable-next-line @typescript-eslint/explicit-function-return-type
 function mkFrame() {
