+++ src/internal/attach.ts
