+++ src/internal/constants.ts
