+++ src/internal/has-opener.ts
