+++ src/internal/is-not-string.ts
