+++ src/internal/message.ts
