+++ src/methods/on.ts
@@ -1,7 +1,10 @@
-import type { FramebusOnHandler, FramebusSubscribeHandler } from "../lib/types";
+import type {
+  FramebusOnHandler,
+  FramebusSubscribeHandler,
+} from "../internal/types";
 import type { FramebusConfig } from "../framebus-config";
-import { subscriptionArgsInvalid } from "../lib/subscription-args-invalid";
-import { subscribers } from "../lib/constants";
+import { subscriptionArgsInvalid } from "../internal/subscription-args-invalid";
+import { subscribers } from "../internal/constants";
 
 export function on(
   config: FramebusConfig,
